moving exec processing to its own package so it can be reused outside of the engine
